+++ src/renderer/services/chain/transaction/transfer.ts
@@ -1,17 +1,13 @@
 import * as RD from '@devexperts/remote-data-ts'
-import { getTokenAddress } from '@xchainjs/xchain-ethereum'
 import * as FP from 'fp-ts/lib/function'
-import * as O from 'fp-ts/lib/Option'
 import * as Rx from 'rxjs'
 import * as RxOp from 'rxjs/operators'
 
-import { isEthAsset } from '../../../helpers/assetHelper'
-import { isEthChain } from '../../../helpers/chainHelper'
 import { liveData } from '../../../helpers/rx/liveData'
 import { observableState } from '../../../helpers/stateHelper'
 import { INITIAL_SEND_STATE } from '../const'
 import { SendTxStateHandler, SendTxState } from '../types'
-import { poolTxStatusByChain$, sendTx$ } from './common'
+import { sendTx$ } from './common'
 
 /**
  * Send TX
@@ -25,32 +21,13 @@ export const transfer$: SendTxStateHandler = (params) => {
   } = observableState<SendTxState>({
     ...INITIAL_SEND_STATE,
     status: RD.pending,
-    steps: { current: 1, total: 2 }
+    steps: { current: 1, total: 1 }
   })
 
-  // All requests will be done in a sequence
-  // to update `SendTxState` step by step
-  const requests$ = sendTx$(params).pipe(
-    liveData.chain((txHash) => {
-      // Update state
-      setState({
-        ...getState(),
-        steps: { current: 2, total: 2 }
-      })
-      // 2. check tx finality by polling its tx data
-      const erc20Address = FP.pipe(
-        params.asset,
-        // ETH chain only
-        O.fromPredicate(({ chain }) => isEthChain(chain)),
-        // ignore ETH asset
-        O.chain(O.fromPredicate(FP.not(isEthAsset))),
-        O.map(getTokenAddress),
-        O.chain(O.fromNullable)
-      )
-      return poolTxStatusByChain$({ txHash, chain: params.asset.chain, assetAddress: erc20Address })
-    }),
+  const requests$ = FP.pipe(
+    sendTx$(params),
     // Update state
-    liveData.map(({ hash }) => setState({ ...getState(), status: RD.success(hash) })),
+    liveData.map((txHash) => setState({ ...getState(), status: RD.success(txHash) })),
     // Add failures to state
     liveData.mapLeft((apiError) => {
       setState({ ...getState(), status: RD.failure(apiError) })
@@ -63,5 +40,8 @@ export const transfer$: SendTxStateHandler = (params) => {
     })
   )
 
-  return Rx.combineLatest([getState$, requests$]).pipe(RxOp.switchMap(() => Rx.of(getState())))
+  return FP.pipe(
+    Rx.combineLatest([getState$, requests$]),
+    RxOp.switchMap(() => Rx.of(getState()))
+  )
 }