+++ src/renderer/views/wallet/Interact/BondView.tsx
@@ -104,7 +104,7 @@ export const BondView: React.FC<Props> = ({ walletType, walletIndex, walletAddre
         </Styled.ErrorView>
       ),
       (txHash) => (
-        <Styled.SuccessView title={intl.formatMessage({ id: 'common.success' })}>
+        <Styled.SuccessView title={intl.formatMessage({ id: 'common.tx.success' })}>
           <Styled.ViewTxButton txHash={O.some(txHash)} onClick={goToTransaction} />
           <Button onClick={resetInteractState}>{intl.formatMessage({ id: 'common.back' })}</Button>
         </Styled.SuccessView>