+++ src/renderer/views/wallet/Interact/UnbondView.tsx
@@ -79,7 +79,7 @@ export const UnbondView: React.FC<Props> = ({ walletType, walletIndex, openExplo
         </Styled.ErrorView>
       ),
       (txHash) => (
-        <Styled.SuccessView title={intl.formatMessage({ id: 'common.success' })}>
+        <Styled.SuccessView title={intl.formatMessage({ id: 'common.tx.success' })}>
           <Styled.ViewTxButton txHash={O.some(txHash)} onClick={goToTransaction} />
           <Button onClick={resetInteractState}>{intl.formatMessage({ id: 'common.back' })}</Button>
         </Styled.SuccessView>